use correct types

+++ src/friendlist.c
@@ -105,7 +105,7 @@ static void friendlist_onMessage(ToxWindow *self, Tox *m, int32_t num, uint8_t *
             str[len] = '\0';
 
             uint8_t nick[TOX_MAX_NAME_LENGTH] = {'\0'};
-            uint16_t n_len = tox_get_name(m, num, nick);
+            int n_len = tox_get_name(m, num, nick);
             n_len = MIN(n_len, TOXIC_MAX_NAME_LENGTH);
             nick[n_len] = '\0';
 
@@ -215,7 +215,7 @@ static void friendlist_onFileSendRequest(ToxWindow *self, Tox *m, int32_t num, u
             tox_file_send_control(m, num, 1, filenum, TOX_FILECONTROL_KILL, 0, 0);
 
             uint8_t nick[TOX_MAX_NAME_LENGTH];
-            uint16_t n_len = tox_get_name(m, num, nick);
+            int n_len = tox_get_name(m, num, nick);
             n_len = MIN(n_len, TOXIC_MAX_NAME_LENGTH);
             nick[n_len] = '\0';
 
@@ -238,7 +238,7 @@ static void friendlist_onGroupInvite(ToxWindow *self, Tox *m, int32_t num, uint8
             friends[num].chatwin = add_window(m, new_chat(m, friends[num].num));
         } else {
             uint8_t nick[TOX_MAX_NAME_LENGTH];
-            uint16_t n_len = tox_get_name(m, num, nick);
+            int n_len = tox_get_name(m, num, nick);
             n_len = MIN(n_len, TOXIC_MAX_NAME_LENGTH);
             nick[n_len] = '\0';
 
@@ -545,7 +545,7 @@ static void friendlist_onAv(ToxWindow *self, ToxAv *av)
             friends[id].chatwin = add_window(m, new_chat(m, friends[id].num));
         } else {
             uint8_t nick[TOX_MAX_NAME_LENGTH] = {'\0'};
-            uint16_t n_len = tox_get_name(m, id, nick);
+            int n_len = tox_get_name(m, id, nick);
 
             n_len = MIN(n_len, TOXIC_MAX_NAME_LENGTH);
             nick[n_len] = '\0';
